+++ RELEASE-NOTES.md
@@ -10,6 +10,11 @@
   - New getter/setter & `Init` functions
   - `NewContext` now returns single `context.Context` result
   - New functions `NewContextIfNil` and `NewContextWithCancel`
+- `SqlConnect`:
+  - New getter/setter & `Init` functions
+  - New db flavor for SQLite
+  - `NewContext` now returns single `context.Context` result
+  - New functions `NewContextIfNil` and `NewContextWithCancel`
 - More unit tests.
 - Update dependency libs.
 - Other fixes & enhancements.